move placeholder-dump into a separate module
It's too heavyweight to load the whole chaintool app for each completion attempt, and the dump code clutters up the app anyway.

As part of this refactor, sequence_impl has been split into op and core modules.

Also note that here the CHAINTOOL_SHORTCUT_PYTHON variable optionally used to force the python version for running chaintool in shortcuts has been renamed CHAINTOOL_PYTHON. It's now used both in shortcuts and also in the completion script when running this placeholders-dump module.